+++ parsnip/parsnip.py
@@ -542,9 +542,9 @@ class CifFile:
 
     def build_unit_cell(
         self,
-        n_decimal_places: int = 4,
+        n_decimal_places: int = 3,
         additional_columns: str | Iterable[str] | None = None,
-        parse_mode: Literal["python_float", "sympy"] = "python_float",
+        parse_mode: Literal["rational", "python_float", "sympy"] = "rational",
         verbose: bool = False,
     ):
         """Reconstruct fractional atomic positions from Wyckoff sites and symops.
@@ -561,8 +561,17 @@ class CifFile:
             If the parsed unit cell has more atoms than expected, decrease
             ``n_decimal_places`` to account for noise. If the unit cell has fewer atoms
             than expected, increase ``n_decimal_places`` to ensure atoms are compared
-            with sufficient precision. In many cases, setting ``parse_mode='sympy'``
-            can improve the accuracy of reconstructed unit cells.
+            with sufficient precision.
+
+        .. tip::
+
+            For improved performance, install the `cfractions`_ library. ``parsnip``
+            will automatically detect if it is installed and use the more optimized path
+            when constructing unit cells.
+
+
+        .. _cfractions: https://pypi.org/project/cfractions/
+
 
         Example
         -------
@@ -603,10 +612,12 @@ class CifFile:
                 the Wyckoff site positions. This data is replicated alongside the atomic
                 coordinates and returned in an auxiliary array.
                 Default value = ``None``
-            parse_mode : {'sympy', 'python_float'}, optional
-                Whether to parse lattice sites symbolically (``parse_mode='sympy'``) or
-                numerically (``parse_mode='python_float'``). Sympy is typically more
-                accurate, but may be slower. Default value = ``'python_float'``
+            parse_mode : {'rational', 'sympy', 'python_float'}, optional
+                Whether to parse lattice sites using rational
+                (``parse_mode='rational'``) or floating-point
+                (``parse_mode='python_float'``) arithmetic. 'rational' is more accurate
+                than 'python_float', but may take more time.
+                Default value = ``'rational'``
             verbose : bool, optional
                 Whether to print debug information about the uniqueness checks.
                 Default value = ``False``
@@ -631,7 +642,15 @@ class CifFile:
                 "Sympy is not available! Please set parse_mode='python_float' "
                 "or install sympy."
             )
-        valid_modes = {"sympy", "python_float"}
+        if parse_mode == "sympy":
+            msg = (
+                "The `sympy` parse mode is deprecated in favor of `rational`, the new "
+                "default. `rational` is as accurate as sympy, much faster, and does not"
+                " require any dependencies. Please convert existing codes to use the "
+                "`parse_mode='rational'`."
+            )
+            warnings.warn(msg, category=DeprecationWarning, stacklevel=2)
+        valid_modes = {"rational", "sympy", "python_float"}
         if parse_mode not in valid_modes:
             raise ValueError(f"Parse mode '{parse_mode}' not in {valid_modes}.")
 
@@ -664,11 +683,16 @@ class CifFile:
         )
 
         frac_strs = self._read_wyckoff_positions()
+        if len(frac_strs) == 0:
+            msg = (
+                "No Wyckoff positions were found when constructing unit cell. "
+                f"Found wyckoff_keys: {self._raw_wyckoff_keys or None}"
+            )
+            raise ParseError(msg)
 
         all_frac_positions = [
             _safe_eval(symops_str, *xyz, parse_mode=parse_mode) for xyz in frac_strs
         ]  # Compute N_symmetry_elements coordinates for each Wyckoff site
-
         pos = np.vstack(all_frac_positions)
 
         # Wrap into box - works generally because these are fractional coordinates
@@ -845,8 +869,15 @@ class CifFile:
         wyckoff_position_data = [
             self.get_from_loops(key) for key in self.__class__._WYCKOFF_KEYS
         ]
+        wyckoff_position_data = [
+            (col[(col != ".") & (col != "?")][:, None] if col is not None else col)
+            for col in wyckoff_position_data
+        ]
 
-        if all(x is None for x in wyckoff_position_data) and self._strict:
+        if (
+            all((x is None or len(x) == 0) for x in wyckoff_position_data)
+            and self._strict
+        ):
             msg = "No wyckoff position data was found!"
             raise ParseError(msg)
 
@@ -1010,6 +1041,7 @@ class CifFile:
 
                 if n_elements % n_cols != 0:
                     msg = (
+                        f"CifFile('{self._fn}') : "
                         f"Parsed data for table {len(self.loops) + 1} cannot be"
                         f" resolved into a table of the expected size and will be "
                         f"ignored. \nGot n={n_elements} items, which cannot be "