+++ changelog.rst
@@ -4,6 +4,23 @@ Changelog
 The format is based on `Keep a Changelog <http://keepachangelog.com/en/1.1.0/>`__.
 This project adheres to `Semantic Versioning <http://semver.org/spec/v2.0.0.html>`__.
 
+v0.X.X - 20XX-XX-XX
+-------------------
+
+Added
+~~~~~
+- New rational parsing mode based on ``fractions.Fraction`` (#161)
+- Spacegroup symop lookup when a table of translation operators is not provided (#160)
+
+Changed
+~~~~~~~
+- ``parse_mode='rational'`` is now the default setting in build_unit_cell (#161)
+- ``n_decimal_places=3`` is now the default setting in build_unit_cell (#161)
+
+Deprecated
+~~~~~~~~~~
+- ``parse_mode='sympy'`` is now deprecated in favor of ``rational`` (#161)
+
 v0.5.0 - 2026-01-02
 -------------------